Hotkeys: minor changes
