+++ client/src/components/app/App/index.tsx
@@ -1,24 +1,16 @@
 import React from 'react';
 import { Router } from '@reach/router';
 import StoreContext from 'storeon/react/context';
-import mousetrap from 'mousetrap';
 import { PrivateRoute, PublicRoute } from '../../Route';
 import Home from '../Home';
 import Login from '../Login';
 import SnackbarProvider from '../../Snackbar';
 import { store } from '../../../store';
 import hotkeys from '../../../hotkeys';
+import useHotkey from '../../../hooks/useHotkey';
 
 function App() {
-  React.useEffect(() => {
-    mousetrap.bind(hotkeys.help, () => {
-      store.dispatch('help/show');
-    });
-
-    return () => {
-      mousetrap.unbind(hotkeys.help);
-    }
-  }, []);
+  useHotkey(hotkeys.help, () => store.dispatch('help/show'));
 
   return (
     <StoreContext.Provider value={store}>
