+++ client/src/components/ui/Button/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import styles from './styles.module.scss';
-import classes from '../../../utils/classes';
 
 type Props = {
   children: any;
@@ -11,7 +10,7 @@ type Props = {
 
 export default function Button({ children, onClick = () => {}, type = 'submit', ...props }: Props) {
   return (
-    <button onClick={onClick} type={type} className={classes("ripple", styles.button)} {...props}>
+    <button onClick={onClick} type={type} className={styles.button} {...props}>
       {children}
     </button>
   );
