+++ client/src/components/ui/HotkeyHelp/styles.module.scss
@@ -15,6 +15,7 @@
   padding: 1px 3px 5px;
   pointer-events: none;
   position: absolute;
+  text-transform: none;
   top: -6px;
   vertical-align: middle;
   white-space: nowrap;
