+++ client/src/index.scss
@@ -1,5 +1,4 @@
 @import './styles/_fonts';
-@import './styles/_ripple';
 
 body {
   margin: 0;
